[v3.0] WebShop and Retailer Application Scenario
This commit contains the sources of the WebShop Application which is
running on SAP HANA Cloud
https://cloudmodelespmhana.hana.ondemand.com/espm-cloud-web/webshop

The corresponding WebShop Scenario description on SCN
Find the link of new ESPM Java WebShop & Retailer
scenario tutorial here
http://scn.sap.com/docs/DOC-46869

This tutorial explains what the ESPM Model is and how it is build in
Java as the basis for the WebShop application.
Furthermore by means of the so caller Retailer applications it is
explained how Roles and Security concepts of HANA Cloud Web applications
are explained.