[v3.0] Step 1: Starting WebShop Extension Scenario This commit contains a project frame which is a good starting point for any SAPUI5 based HANA Cloud Application development. It is used here as starting point for the WebShop Extension Scenario.
It offers a default structure of a typical SAPUI5 Maven
project and by very small UI5 examples included in the frame it is
demonstrated how to read and write data from backend systems by means of
OData Services.

The corresponding WebShop Extension Scenario step by step tutorial on
SCN
http://scn.sap.com/docs/DOC-41491
makes use of this project frame and in in chapter 2 it is described how
the project frame is build up and how to run it.

Used technologies SAPUI5, OData, SAP HANA Cloud Connectivity.