[v3.0] Step 2: 'Select Product' UI Component of the WebShop Extension Scenario
This commit contains the sources of the 'Select Product' UI component
of the WebShop Extension Scenario. This component is part of the WebShop
Extension Application ('Customer Reviews') which is running as demo
application on SAP HANA Cloud
https://webreviewsespmhana.hana.ondemand.com/espm-ui-reviews-web/

The corresponding WebShop Extension Scenario description on SCN
http://scn.sap.com/docs/DOC-41491
is a step by step tutorial for any developers. 
In Chapter 2 it is described how the 'Select Product' UI component is
build up and how you run it on SAP HANA Cloud server.

Used technologies SAPUI5, OData, SAP HANA Cloud Connectivity