[v3.0] Step 4: 'Review Creation' UI Component of WebShop Extension Scenario
This commit contains the sources of the 'Review Creation' UI component
of the WebShop Extension Scenario. This component is part of the WebShop
Extension Application ('Customer Reviews') which is running as demo
application on SAP HANA Cloud
https://webreviewsespmhana.hana.ondemand.com/espm-ui-reviews-web/

The corresponding WebShop Extension Scenario description on SCN
http://scn.sap.com/docs/DOC-41491
is a step by step tutorial for any developer. 
In Chapter 4 it is described how the 'Review Creation' UI component is
build up and how it runs.

Used technologies SAPUI5, OData, SAP HANA Cloud Connectivity